refactor(core): split ApplicationService into AuditService + public SecurityHeaders
The application->core module rename left ApplicationService reading as
core.ApplicationService while holding two unrelated helpers. Split them to
their proper homes: record_audit moves to AuditService (fronting AuditWriter),
and SecurityHeaders becomes a public core class that server.py calls directly,
removing the passthrough. ApplicationService is deleted. ApplicationRepository
and ApplicationRepositoryClient keep their names.
