+++ src/apps/backend/modules/core/audit_service.py
@@ -1,17 +1,10 @@
 from typing import Optional
 
-from flask import Flask
-
 from modules.core.common.types import AuditActor, AuditOutcome, FieldChanges, ResourceAction
 from modules.core.internal.audit.audit_writer import AuditWriter
-from modules.core.internal.security_headers import SecurityHeaders
 
 
-class ApplicationService:
-    @staticmethod
-    def install_security_headers(app: Flask) -> None:
-        SecurityHeaders.init_app(app)
-
+class AuditService:
     @staticmethod
     def record_audit(
         *,
@@ -22,8 +15,6 @@ class ApplicationService:
         changes: Optional[FieldChanges] = None,
         outcome: AuditOutcome = AuditOutcome.SUCCESS,
     ) -> None:
-        # Rarely needed: ApplicationRepository already audits every create/update/delete. Use this only
-        # for an access a custom reader/writer performs that the generic CRUD does not cover.
         AuditWriter.record(
             actor=actor,
             resource_type=resource_type,
