+++ src/apps/backend/modules/core/security_headers.py
